Add profile load and import feature ; with command 'new profile'

+++ palang-virtual-machine/src/llm/model_settings.rs
@@ -1,5 +1,5 @@
 pub struct ModelSettings {
     pub model: String,
-    pub temperature: u32,
+    pub temperature: f32,
     pub max_tokens: u32,
 }
